requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
aiosqlite
python-multipart
aiofiles
pyjwt
websockets
greenlet